src/requirements/requirements.txt
flask==2.0.1
flask-cors==4.0.0
flask-restx==1.1.0
gunicorn==21.2.0
python-dotenv==1.0.0
xmltodict==0.13.0
werkzeug==2.0.3
kubernetes==29.0.0
aiohttp==3.9.1
asyncio==3.4.3
prometheus-client==0.17.1
psutil==5.9.6

# Testing Dependencies
pytest>=7.4.0
pytest-mock>=3.11.1

# Metrics Dependencies
matplotlib>=3.7.0
numpy>=1.24.0
